Move the Type enum to its own file
This is a good start to cleaning up common.h.
